requirements.txt
numpy
matplotlib
scikit-image
rawpy
imageio.v3